refactor(cache): decompose CacheImageResourceOperation into collaborators
The 768-line operation mixed validation setup, upstream fetching with
negative caching and streaming size guards, SVG sanitization, Sharp
processing, default-image fallback, cache/filesystem reads, and atomic
writes behind 9 constructor dependencies. Behaviour is moved verbatim:

- ResourceFetcher: negative cache, upstream fetch, content-length +
  streaming size enforcement, store-to-temp-file
- ImageFormatProcessor: SVG header detection/sanitization, raster
  processing, default-image pipeline, metadata construction (previously
  repeated four times)
- svg-sanitizer.util: pure sanitizeSvg() extracted from the class
- CacheImageResourceOperation keeps its public API (setup, execute,
  checkResourceExists, fetchHeaders, getCachedResource, path getters,
  optimizeAndServeDefaultImage) and now orchestrates; the six repeated
  endPhase+recordCacheOperation blocks collapse into one helper

Verified end-to-end: production boot serves a valid WebP through the
fetch->fallback->process->stream path; full suite and e2e green.
